Use 1 hour step size
rrdtool otherwise calculates valuesu that are probably correct with the
way rrdtool works, but not entirely what I want. If we don't receive any
messages over the rest of the heartbeat period, we want rrdtool to
extrapolate up to the next RRA definition.

A script, copied in to the end of this commit, has been written to
verify the change. It uses timestamps to try and replicate the
behaviour. rrdtool will generate these values with a 15 minute step
size:

    -- RRDTOOL --
    2023-04-27 20:00:00 37.4048443332
    2023-04-27 21:00:00 43.970222538
    2023-04-27 22:00:00 59.52309914519999
    -- PYTHON -- (H-1)
    20230427 19:00:00 29
    20230427 20:00:00 23
    20230427 21:00:00 32

The Python script (please add 1 to the hour) simply sums up the messages
in a timestamp. rrdtool, with step size < 1 hour, seems to extrapolate
the data we gave it. As we only send in data when a message arrives and
not for the case of no messages, this is not fully what we want: we want
it to utilize the lack of information as the information that no more
messages were received during that time. With it matching the step size,
we have intervals in the RRA with UNKNOWN values and intervals with the
proper values. In theory we could make the step higher, but we can
simply use the `ADDNAN` RPN function to change those to 0.

With the new step size, it looks as follows:

    -- RRDTOOL --
    2023-04-27 20:00:00 37.4048443332
    2023-04-27 21:00:00 22.98830222736
    2023-04-27 22:00:00 32.110548894
    -- PYTHON -- (H-1)
    20230427 19:00:00 29
    20230427 20:00:00 23
    20230427 21:00:00 32

Much better. The following Python script can be used to play around with
it:

    import datetime
    import math
    import os
    import os.path
    import subprocess

    # rrd = actual, 24 = 24, 37 = 29, 43 = 23

    data = (
        1682615695.827, 1682616255.775, 1682616259.824, 1682616264.214, 1682616269.964, 1682616971.314, 1682616975.763, 1682616978.494, 1682616986.305, 1682617012.893, 1682617021.708, 1682617025.928, 1682617047.302, 1682617090.121, 1682617094.576, 1682617103.667, 1682617107.128, 1682617154.64, 1682617161.273, 1682617165.826, 1682617658.291, 1682617675.112, 1682617689.207, 1682618227.746, 1682618232.291, 1682618297.081, 1682618363.3, 1682618370.587, 1682618385.284, 1682618627.416, 1682619025.345, 1682619072.886, 1682619076.094, 1682619084.842, 1682620727.627, 1682620766.105, 1682620786.588, 1682620788.665, 1682620790.384, 1682620899.23, 1682620913.612, 1682620960.385, 1682621323.186, 1682621682.794, 1682621688.633, 1682621691.544, 1682621700.866, 1682621714.378, 1682621722.653, 1682621930.071, 1682621963.379, 1682621964.629, 1682622677.765, 1682622950.395, 1682622951.681, 1682622955.675, 1682622958.851, 1682622961.887, 1682622965.052, 1682622965.078, 1682623055.465, 1682623058.592, 1682623084.059, 1682623092.174, 1682623097.459, 1682623156.748, 1682623160.971, 1682623167.926, 1682623168.854, 1682623207.527, 1682623222.665, 1682623261.971, 1682624724.759, 1682624740.426, 1682625116.583, 1682625121.557, 1682625142.121, 1682625392.932, 1682625535.442, 1682625562.031, 1682625576.621, 1682625591.091, 1682625593.391, 1682625593.96, 1682625631.798, 1682625644.734,
    )

    def discordtounix(ts):
        return ((ts >> 22) + 1420070400000) / 1000

    if os.path.exists("test.rrd"):
        os.remove("test.rrd")

    subprocess.check_output([
        "rrdtool", "create", "--start", str(int(data[0])), "test.rrd", "--step", "5m",
        "--no-overwrite", "DS:messages:ABSOLUTE:15m:U:U",
        "RRA:AVERAGE:0.5:1h:14d", "RRA:AVERAGE:0.5:1d:6M", f"RRA:AVERAGE:0.5:1w:{52 * 5}w"
    ])

    for item in sorted(data):
        try:
            subprocess.check_output([
                "rrdtool", "update", "test.rrd", str(item) + ":1"])
        except subprocess.CalledProcessError as exc:
            print(exc.returncode, exc.output)
            exit(1)

    proc = subprocess.check_output("rrdtool fetch test.rrd -a AVERAGE".split(), universal_newlines=True)
    for line in proc.splitlines():
        if ':' in line:
            ts, val = line.split(': ')
            parsedts = datetime.datetime.fromtimestamp(int(ts))
            parsedval = float(val)
            if not math.isnan(parsedval):
                print(parsedts, parsedval * 60 * 60)

    # vim: set nowrap: